T-Head separatrix probe: independent confirmation of v8a MAH interpretation (a)
Ran scripts/thead_eval.py on the same 61-item separatrix battery used for the
v8a probe (commit e1c8dcf), reading frozen Qwen2.5-7B hidden states at L16
through a 131K-param Takens delay-embedding head. No SRT-Adapter loaded; no
parameter sharing with MAH; no training on this probe.

Calibration: swept recurrence_eps in {1,3,10,15,20,25,28,30,100}; below 15
the head reports zero recurrence, at 25+ it saturates. eps=20 hits the
transition zone (rec_mean ~3.6 / max ~21 / T ~30 -> ~10%). Used eps=20.

Headline (Wilcoxon signed-rank, two-sided):

  Lyapunov std   mystic - bedrock = +0.327  p=2e-9   frac+=0.89
  Lyapunov std   tech   - mystic  = -0.263  p=4e-10  frac+=0.08
    => mystical branch is the MOST trajectory-volatile

  Recurrence mean   bedrock - tech    = +3.17  p=1e-6
  Recurrence mean   bedrock - mystic  = +2.77  p=2e-5
  Recurrence max    bedrock - mystic  = +6.36  p=7e-8
    => bedrock branch is the MOST recurrent (structured-attractor) trajectory

The T-Head and the SRT-Adapter MAH agree on the rank ordering of the three
branches: bedrock most distinctively-structured (high MAH peak, high
recurrence), mystical most volatile (high Lyapunov std), technical between.
This is independent confirmation of MAH falsification interpretation (a)
'bedrock is the harder collision' — see docs/SEPARATRIX_V8A_FINDINGS.md
for the original three candidate readings.

Two parameter-disjoint heads with no shared theoretical motivation
producing the same rank ordering on the same probe is a stronger
constraint than either head alone. This is also the first direct empirical
instance of the SRT / Geofinitism program convergence sketched in §6.8.

Adds:
  scripts/thead_eval.py     (fp32 cdist fix; T-Head kept in fp32 over bf16 backbone)
  artifacts/thead/separatrix_v1/readouts.jsonl   (207 rows: 3 branches x 61 items + a few prompt_only that survived the min-length filter)
  docs/THEAD_SEPARATRIX_V1_FINDINGS.md           (writeup + interpretation)

Updates paper.md §6.9 with a new paragraph 'Independent confirmation from a
Takens-based readout' documenting the cross-program convergence.

+++ scripts/thead_eval.py
@@ -157,7 +157,7 @@ def main() -> int:
         recurrence_eps=args.recurrence_eps,
         project_to=args.project_to,
     )
-    thead = TakensHead(thead_cfg, d_hidden=d_hidden).to(device).to(load_dtype)
+    thead = TakensHead(thead_cfg, d_hidden=d_hidden).to(device).to(torch.float32)
     thead.eval()
     log.info("T-Head config=%s params=%d", thead_cfg,
              sum(p.numel() for p in thead.parameters()))
@@ -191,7 +191,7 @@ def main() -> int:
                 # hidden_states: tuple of length n_layers+1 (incl. embeddings)
                 h = out.hidden_states[layer_idx + 1]  # +1 to skip embed layer
                 # truncate to real tokens; T-Head ops are O(T^2) so keep tight
-                h = h[:, :T, :]
+                h = h[:, :T, :].float()  # T-Head runs fp32 (cdist needs fp32 on CUDA)
 
                 t_out = thead(h)
                 lyap = t_out.lyapunov[0].float().cpu()
